Consolidate metadata helper functions in metadata.go
